+++ src/onegov.wtfs/onegov/wtfs/tests/test_security.py
@@ -249,13 +249,16 @@ def test_permissions(wtfs_app, wtfs_password):
     for user in (admin, admin_a, admin_b):
         for model in (admin, admin_a, admin_b):
             assert permits(user, model, Public)
-            assert not permits(user, model, AddModel)
-            assert not permits(user, model, AddModelUnrestricted)
-            assert not permits(user, model, EditModel)
-            assert not permits(user, model, EditModelUnrestricted)
-            assert not permits(user, model, DeleteModel)
-            assert not permits(user, model, ViewModel)
-            assert not permits(user, model, ViewModelUnrestricted)
+            assert permits(user, model, AddModel)
+            assert permits(user, model, AddModelUnrestricted)
+            assert permits(user, model, EditModel)
+            assert permits(user, model, EditModelUnrestricted)
+            if user == model:
+                assert not permits(user, model, DeleteModel)
+            else:
+                assert permits(user, model, DeleteModel)
+            assert permits(user, model, ViewModel)
+            assert permits(user, model, ViewModelUnrestricted)
         for model in (editor, editor_a, editor_b, member, member_a, member_b):
             assert permits(user, model, Public)
             assert permits(user, model, AddModel)